fix(bluedbm): do not use the bitmap in the end request
I found that using the bitmap in the end request may occur the serious
problem in the execute based on bluedbm platform. Therefore, I changed
it to the atomic global array.

`gint` is generally treated as 8 bytes. In other words, with the 4096
segments, it only needs the 8 * 4096 bytes to handle the segment is bad
or not. In short, that is negligible.