feat(pgftl): add the optional global rwlock
I added the optional global rwlock, this is legacy lock for the page
ftl. Therefore, I am not recommend to use this lock. However, in the ZNS
device, this needs to sequential write in that device. For this reason,
I added this function.

Also, I made test file more selective.